backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.0
